+++ gaios/OnBoard/Controllers/WatchOnlyViewController.swift
@@ -150,7 +150,7 @@ class WatchOnlyViewController: KeyboardViewController {
         }.catch { error in
             session.disconnect()
             switch error {
-            case AuthenticationTypeHandler.AuthError.ConnectionFailed:
+            case LoginError.connectionFailed:
                 DropAlert().error(message: NSLocalizedString("id_connection_failed", comment: ""))
             default:
                 DropAlert().error(message: NSLocalizedString("id_login_failed", comment: ""))